wordcounter/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-rest-framework==0.1.0
django-cors-headers==3.10.1
